\name{NthChina}
\alias{NthChina}
\title{Historical Earthquakes of North China}
\description{
Contains 65 large historical earthquakes in North China between 1480 and 1997, as given by Bebbington & Harte (2003). Events are divided into 4 regions using the regionalisations given by Zheng & Vere-Jones (1991).
}
\usage{data(NthChina)
}
\format{
A data frame with 65 rows, each representing an earthquake event, with the following variables:
\describe{
\item{time}{number of years since 1480 AD.}
\item{latitude}{number of degrees north.}
\item{longitude}{number of degrees east.}
\item{magnitude}{number of magnitude units \emph{above} 6.}
\item{region}{1, 2, 3, or 4; being the region of the event.}
}}

\keyword{datasets}
